refactor(commands): extract test execution into commands/test.rs (Issue #186)
Move the test discovery/execution machinery (~1076 lines) out of the
6841-line commands.rs into commands/test.rs as the first incremental
step of the per-command module split. No behavior change - verified
via cargo build, clippy, fmt, and the full test suite.
